Development - Added `test.sql` file for unit-test tables. - Added more unit-tests. - Optimisations
